requirements.txt
openai
langchain
pinecone-client
dotenv
flask